Make StatusService own "pluggedInAt"
The main activity can't own pluggedInAt, because it generally isn't
alive to receive plugged-in and unplugged events. The StatusService,
however, is long lived, and can receive these events. Move ownership of
this data to StatusService, and share it via SharedPreferences.

This is not ideal, but avoids needing to design a whole IPC protocol for
sharing data between StatusService and the main activity for the time
being.

See: https://github.com/dubrowgn/wattz/issues/10