Make StatusService own all battery data
Since dynamic current units detection was added, the system battery
manager no longer owned all of the data used by both StatusService and
MainActivity. This opened the possibility for the two classes to display
different values for e.g. power. Instead, make StatusService own all
battery data, and use pub/sub to communicate this data with
MainActivity. This restores a single source of truth for battery data,
and removes the possibility for the two classes to disagree.